reorganized and fixed tests
