+++ src/__tests__/useFetch.test.tsx
@@ -9,6 +9,7 @@ import { toCamel } from 'convert-keys'
 import { renderHook, act } from '@testing-library/react-hooks'
 import mockConsole from 'jest-mock-console'
 import * as mockdate from 'mockdate'
+import defaults from '../defaults'
 
 import { Res, Options, CachePolicies } from '../types'
 import { emptyCustomResponse, sleep, makeError } from '../utils'
@@ -92,9 +93,9 @@ describe('useFetch - BROWSER - basic functionality', (): void => {
       var formData = new FormData()
       formData.append('username', 'AlexCory')
       await result.current.post(formData)
-      const options = fetch.mock.calls[0][1] || {}
+      const options = fetch.mock.calls[0][1] || { headers: {} }
       expect(options.method).toBe('POST')
-      expect(options.headers).toBeUndefined()
+      expect('Content-Type' in (options as any).headers).toBe(false)
     })
   })
 })
@@ -591,8 +592,9 @@ describe('useFetch - BROWSER - Overwrite Global Options set in Provider', (): vo
   })
 
   it('should only add Content-Type: application/json for POST and PUT by default', async (): Promise<void> => {
-    const expectedHeadersGET = providerHeaders
+    const expectedHeadersGET = { ...defaults.headers, ...providerHeaders }
     const expectedHeadersPOSTandPUT = {
+      ...defaults.headers,
       ...providerHeaders,
       'Content-Type': 'application/json'
     }
@@ -613,7 +615,10 @@ describe('useFetch - BROWSER - Overwrite Global Options set in Provider', (): vo
   })
 
   it('should have the correct headers set in the options set in the Provider', async (): Promise<void> => {
-    const expectedHeaders = providerHeaders
+    const expectedHeaders = {
+      ...defaults.headers,
+      ...providerHeaders
+    }
     const { result } = renderHook(
       () => useFetch(),
       { wrapper }
@@ -625,7 +630,7 @@ describe('useFetch - BROWSER - Overwrite Global Options set in Provider', (): vo
   })
 
   it('should overwrite url and options set in the Provider', async (): Promise<void> => {
-    const expectedHeaders = undefined
+    const expectedHeaders = defaults.headers
     const expectedURL = 'https://example2.com'
     const { result, waitForNextUpdate } = renderHook(
       () => useFetch(expectedURL, globalOptions => {
@@ -644,7 +649,7 @@ describe('useFetch - BROWSER - Overwrite Global Options set in Provider', (): vo
   })
 
   it('should overwrite options set in the Provider', async (): Promise<void> => {
-    const expectedHeaders = undefined
+    const expectedHeaders = defaults.headers
     const { result, waitForNextUpdate } = renderHook(
       () => useFetch(globalOptions => {
         // TODO: fix the generics here so it knows when a header
